Refactor: move MemberController from controller.command to controller
